requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
firebase-admin>=6.4.0
python-dotenv>=1.0.0
pydantic[email]>=2.10.0

